Initial work on new explainer
