+++ glide.lock
@@ -1,64 +1,64 @@
-hash: 5d8cd28f14a6d894f9d35ae4f4aa7e273f6e0853a51d12d3af92e49e5450109d
-updated: 2017-01-27T08:48:49.889775149-05:00
+hash: c83ff795ee2fa2a591063739a4ab70eecc00715f2e40473daf6d15673f06c59f
+updated: 2017-02-04T13:14:35.255112245-05:00
 imports:
 - name: github.com/aws/aws-sdk-go
   version: 1355b456f6ba4a8453249e0dc7743f16ccac362d
   subpackages:
   - aws
   - aws/awserr
-  - aws/awsutil
-  - aws/client
-  - aws/client/metadata
-  - aws/corehandlers
   - aws/credentials
-  - aws/credentials/ec2rolecreds
-  - aws/credentials/endpointcreds
-  - aws/credentials/stscreds
-  - aws/defaults
-  - aws/ec2metadata
-  - aws/endpoints
-  - aws/request
   - aws/session
-  - aws/signer/v4
-  - private/protocol
-  - private/protocol/ec2query
-  - private/protocol/json/jsonutil
-  - private/protocol/jsonrpc
-  - private/protocol/query
-  - private/protocol/query/queryutil
-  - private/protocol/rest
-  - private/protocol/restjson
-  - private/protocol/restxml
-  - private/protocol/xml/xmlutil
-  - private/waiter
   - service/ec2
   - service/efs
   - service/route53
+  - aws/awsutil
   - service/s3
+  - aws/endpoints
+  - aws/client
+  - aws/corehandlers
+  - aws/credentials/stscreds
+  - aws/defaults
+  - aws/request
+  - aws/client/metadata
+  - aws/signer/v4
+  - private/protocol
+  - private/protocol/ec2query
+  - private/waiter
+  - private/protocol/restjson
+  - private/protocol/restxml
   - service/sts
+  - aws/credentials/ec2rolecreds
+  - aws/credentials/endpointcreds
+  - aws/ec2metadata
+  - private/protocol/rest
+  - private/protocol/query/queryutil
+  - private/protocol/xml/xmlutil
+  - private/protocol/jsonrpc
+  - private/protocol/query
+  - private/protocol/json/jsonutil
 - name: github.com/cloudflare/cfssl
   version: db0d0650b6496bfe8061ec56a92edd32d8e75c30
   subpackages:
-  - api
-  - api/client
-  - auth
-  - certdb
-  - cli
+  - csr
   - cli/genkey
   - config
-  - crypto/pkcs7
-  - csr
-  - errors
   - helpers
-  - helpers/derhelpers
-  - info
   - initca
   - log
-  - ocsp/config
   - signer
   - signer/local
-  - signer/remote
+  - errors
+  - cli
+  - auth
+  - ocsp/config
+  - crypto/pkcs7
+  - helpers/derhelpers
+  - certdb
+  - info
   - signer/universal
+  - signer/remote
+  - api/client
+  - api
 - name: github.com/cpuguy83/go-md2man
   version: a65d4d2de4d5f7c74868dfa9b202a3c8be315aaa
   subpackages:
@@ -72,10 +72,12 @@ imports:
   repo: https://github.com/google/certificate-transparency
   subpackages:
   - go
-  - go/asn1
   - go/client
   - go/x509
+  - go/asn1
   - go/x509/pkix
+- name: github.com/gosuri/uilive
+  version: efb88ccd059957c48f24f9d351d33a0eb00ede41
 - name: github.com/inconshreveable/mousetrap
   version: 76626ae9c91c4f2a10f34cad8ce83ea42c93bb75
 - name: github.com/jmespath/go-jmespath
@@ -98,32 +100,32 @@ imports:
   subpackages:
   - config
   - internal/codelocation
-  - internal/containernode
   - internal/failer
-  - internal/leafnodes
   - internal/remote
-  - internal/spec
-  - internal/specrunner
   - internal/suite
   - internal/testingtproxy
   - internal/writer
   - reporters
   - reporters/stenographer
   - types
+  - internal/containernode
+  - internal/leafnodes
+  - internal/spec
+  - internal/specrunner
 - name: github.com/onsi/gomega
   version: d59fa0ac68bb5dd932ee8d24eed631cdd519efc3
   subpackages:
-  - format
   - internal/assertion
   - internal/asyncassertion
-  - internal/oraclematcher
   - internal/testingtsupport
   - matchers
+  - types
+  - internal/oraclematcher
+  - format
   - matchers/support/goraph/bipartitegraph
   - matchers/support/goraph/edge
   - matchers/support/goraph/node
   - matchers/support/goraph/util
-  - types
 - name: github.com/packethost/packngo
   version: 91a3a4f65e3cd6ee66d919b4e7103ae038ce66ff
 - name: github.com/pkg/browser
@@ -142,10 +144,10 @@ imports:
 - name: golang.org/x/crypto
   version: 1f22c0103821b9390939b6776727195525381532
   subpackages:
-  - curve25519
-  - pkcs12
-  - pkcs12/internal/rc2
   - ssh
+  - pkcs12
+  - curve25519
+  - pkcs12/internal/rc2
 - name: golang.org/x/sys
   version: 8f0908ab3b2457e2e15403d3697c9ef5cb4b57a9
   subpackages:
